Relocate surviving ML/pipeline/qualification code linkedin/ -> core/ (browserless, channel-agnostic engine); generate forward migrations: delete the 3 LinkedIn models, add email schema (chat linkedin_urn->external_id rename, SiteConfig.country_code, Deal.next_follow_up_at, Mailbox imap, Task/Deal state); linkedin kept as internal migrations-only shell for backward-compatible upgrades
